+++ lib/wds.dart
@@ -2,3 +2,6 @@ library;
 
 export 'package:wds_components/wds_components.dart';
 export 'package:wds_foundation/wds_foundation.dart';
+
+/// util export
+export 'src/util/util.dart';