+++ CHANGELOG.md
@@ -1,7 +1,10 @@
+### 0.0.11
+- 메세지 UI (WdsToast, WdsSnackbar, WdsSectionMessage) 공통 컨트롤러 도입
+- WdsHeader.search > widthFactor 수정
+
 ### 0.0.10
 - WdsTextArea > 최소 높이 60px / 최대 높이 320px 대응될 수 있게끔 개선
 - WdsTextArea > 'label' 파라미터 String에서 Text로 형변환 
-- WdsToast > size width infinity하게 설정
 
 ### 0.0.9
 - WdsToolTip > 'message'에서 'label'로 변경